+++ py/tests/tests_unit_mojo_pipeline.py
@@ -21,6 +21,7 @@ that mojo can run without H2O runtime in PySparkling environment
 """
 import sys
 import os
+
 sys.path.insert(0, sys.argv[1])
 os.environ['PYSPARK_PYTHON'] = sys.executable
 import os
@@ -33,7 +34,10 @@ from pyspark.sql import SparkSession
 from pysparkling.ml import H2OMOJOPipelineModel
 from pyspark.ml import Pipeline, PipelineModel
 
-
+##
+## These tests does not start H2O Context on purpose to test running predictions
+## in Spark environment without run-time H2O
+##
 class H2OMojoPipelineTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
@@ -91,14 +95,15 @@ class H2OMojoPipelineTest(unittest.TestCase):
                                               header=True)
         # Create Spark pipeline of single step - mojo pipeline
         pipeline = Pipeline(stages=[mojo])
-        pipeline.write().overwrite().save( "file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline"))
-        loaded_pipeline = Pipeline.load( "file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline"))
+        pipeline.write().overwrite().save("file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline"))
+        loaded_pipeline = Pipeline.load("file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline"))
 
         ## Train the pipeline model
         model = loaded_pipeline.fit(prostate_frame)
 
-        model.write().overwrite().save( "file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline_model"))
-        loaded_model = PipelineModel.load( "file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline_model"))
+        model.write().overwrite().save("file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline_model"))
+        loaded_model = PipelineModel.load(
+            "file://" + os.path.abspath("build/test_dai_pipeline_as_spark_pipeline_model"))
 
         preds = loaded_model.transform(prostate_frame).repartition(1).select(mojo.select_prediction_udf("AGE")).take(5)
